Core/DBLayer: Fix some forgotten alignments

+++ src/server/game/Guilds/Guild.cpp
@@ -2476,7 +2476,7 @@ bool Guild::AddMember(uint64 guid, uint8 rankId)
                 fields[0].GetString(),
                 fields[1].GetUInt8(),
                 fields[2].GetUInt8(),
-                fields[3].GetUInt32(),
+                fields[3].GetUInt16(),
                 fields[4].GetUInt32());
 
             ok = member->CheckStats();
